Andrii's Solution

+++ app/main.py
@@ -15,7 +15,7 @@ class CarWashStation:
             self,
             distance_from_city_center: float,
             clean_power: int,
-            average_rating: int,
+            average_rating: float,
             count_of_ratings: int
     ) -> None:
         self.distance_from_city_center = distance_from_city_center
